Split jobs.py into multiple Python files.
